fix(parser): type-member modifier grammar reports TS1070 for export/in/out and TS1024 for readonly on method/construct signatures (#16291)
tsc's checkGrammarModifiers rejects `export`/`in`/`out` as a modifier on an
interface or type-literal member (TS1070) and `readonly` on a method or
construct signature (TS1024). tsz previously mis-parsed `export` to TS1005,
dropped `in`/`out` silently, and stayed silent on `readonly` method/construct
signatures.

Extend the existing type-member modifier-error branch
(parse_type_member_visibility_modifier_error) to cover export/in/out, reusing
the same property-name / line-break lookahead so `export: T` and
`export(){}`-as-name stay clean. Emit TS1024 in
parse_type_member_property_or_method when a consumed `readonly` resolves to a
method/construct signature, anchored at the readonly keyword to match tsc's
exact code + line + column. Both fire for interface and type-literal members via
the shared parse path.

`import`/`const`/`default` are not type-member modifiers in tsc (parser-recovery
path, TS1005/TS1131), which tsz already matches, and are left untouched.

New parser tests cover both gaps with anchored fingerprints, renamed binders,
each container, the readonly/export/in/out combos, and negative property-name /
method-name controls.

The addition pushed state_declarations.rs over the 2000-line architecture
ceiling, so enum declaration parsing (parse_enum_declaration[_with_modifiers],
parse_enum_declaration_name, parse_enum_members) moves verbatim to the new
state_declarations_enums.rs (1815 lines remaining). Its now-compliant ratchet
entries are removed from the arch-guard allowlists (which only shrink); enum
parsing behaviour is unchanged.
