+++ crates/tsz-parser/src/parser/state_declarations.rs
@@ -1,9 +1,11 @@
-//! Parser state - interface, type alias, enum, module, import/export, and control flow parsing methods
+//! Parser state - interface, type alias, module, import/export, and control flow parsing methods
+//!
+//! Enum declaration parsing lives in `state_declarations_enums.rs`.
 
 use super::state::ParserState;
 use crate::parser::{
     NodeIndex, NodeList,
-    node::{EnumData, EnumMemberData, IdentifierData, ParameterData},
+    node::{IdentifierData, ParameterData},
     syntax_kind_ext,
 };
 use tsz_common::interner::{AstAtom, IdentText};
@@ -423,6 +425,9 @@ impl ParserState {
                 | SyntaxKind::OverrideKeyword
                 | SyntaxKind::AbstractKeyword
                 | SyntaxKind::DeclareKeyword
+                | SyntaxKind::ExportKeyword
+                | SyntaxKind::InKeyword
+                | SyntaxKind::OutKeyword
         ) && !self.look_ahead_is_property_name_after_keyword()
             && !self.look_ahead_has_line_break_after_keyword()
         {
@@ -485,14 +490,19 @@ impl ParserState {
         start_pos: u32,
         in_interface_declaration: bool,
     ) -> NodeIndex {
-        let readonly = if self.is_token(SyntaxKind::ReadonlyKeyword)
+        // Capture the `readonly` keyword span so a TS1024 (readonly on a
+        // method/construct signature) can be anchored at the modifier itself,
+        // matching tsc's `checkGrammarModifiers`.
+        let readonly_span = if self.is_token(SyntaxKind::ReadonlyKeyword)
             && !self.look_ahead_is_property_name_after_keyword()
         {
+            let span = (self.token_pos(), self.token_end());
             self.next_token();
-            true
+            Some(span)
         } else {
-            false
+            None
         };
+        let readonly = readonly_span.is_some();
 
         let Some(name) = self.parse_type_member_property_or_method_name(start_pos, readonly) else {
             return NodeIndex::NONE;
@@ -509,6 +519,18 @@ impl ParserState {
         let modifiers = self.readonly_modifier_node_list(start_pos, readonly);
 
         if self.is_token(SyntaxKind::OpenParenToken) || self.is_token(SyntaxKind::LessThanToken) {
+            // `readonly` is legal only on a property declaration or index
+            // signature; on a method or construct signature tsc reports TS1024,
+            // anchored at the `readonly` keyword.
+            if let Some((ro_start, ro_end)) = readonly_span {
+                use tsz_common::diagnostics::diagnostic_codes;
+                self.parse_error_at(
+                    ro_start,
+                    ro_end.saturating_sub(ro_start),
+                    "'readonly' modifier can only appear on a property declaration or index signature.",
+                    diagnostic_codes::READONLY_MODIFIER_CAN_ONLY_APPEAR_ON_A_PROPERTY_DECLARATION_OR_INDEX_SIGNATURE,
+                );
+            }
             return self.parse_type_member_method_signature(
                 start_pos,
                 name,
@@ -1372,214 +1394,6 @@ impl ParserState {
                 type_node,
             },
         )
-    }
-
-    /// Parse enum declaration
-    pub(crate) fn parse_enum_declaration(&mut self) -> NodeIndex {
-        let start_pos = self.token_pos();
-        self.parse_enum_declaration_with_modifiers(start_pos, None)
-    }
-
-    /// Parse enum declaration with explicit modifiers
-    pub(crate) fn parse_enum_declaration_with_modifiers(
-        &mut self,
-        start_pos: u32,
-        modifiers: Option<NodeList>,
-    ) -> NodeIndex {
-        let enum_keyword_end = self.token_end();
-        self.parse_expected(SyntaxKind::EnumKeyword);
-
-        let name = self.parse_enum_declaration_name();
-
-        let has_open_brace = self.parse_expected(SyntaxKind::OpenBraceToken);
-
-        let members = if has_open_brace {
-            self.parse_enum_members()
-        } else {
-            Self::make_node_list(Vec::new())
-        };
-
-        if has_open_brace {
-            self.parse_expected(SyntaxKind::CloseBraceToken);
-        }
-
-        let end_pos = if has_open_brace {
-            self.token_end()
-        } else {
-            enum_keyword_end
-        };
-        self.arena.add_enum(
-            syntax_kind_ext::ENUM_DECLARATION,
-            start_pos,
-            end_pos,
-            EnumData {
-                modifiers,
-                name,
-                members,
-            },
-        )
-    }
-
-    fn parse_enum_declaration_name(&mut self) -> NodeIndex {
-        let start_pos = self.token_pos();
-        let end_pos = self.token_end();
-
-        if self.is_reserved_word() {
-            // `tsc` reports the missing enum name but leaves the reserved word
-            // for the outer statement parser. This preserves recovered forms like
-            // `enum void {}` as an anonymous enum plus a following `void {}`.
-            self.error_identifier_expected();
-            return self.arena.add_identifier(
-                SyntaxKind::Identifier as u16,
-                start_pos,
-                end_pos,
-                IdentifierData {
-                    atom: AstAtom::NONE,
-                    escaped_text: IdentText::empty(),
-                    original_text: None,
-                },
-            );
-        }
-
-        self.parse_identifier()
-    }
-
-    /// Parse enum members
-    pub(crate) fn parse_enum_members(&mut self) -> NodeList {
-        use tsz_common::diagnostics::diagnostic_codes;
-        let mut members = Vec::new();
-
-        while !self.is_token(SyntaxKind::CloseBraceToken)
-            && !self.is_token(SyntaxKind::EndOfFileToken)
-        {
-            let start_pos = self.token_pos();
-
-            // Handle leading comma - emit TS1132 "Enum member expected" and skip
-            if self.is_token(SyntaxKind::CommaToken) {
-                self.parse_error_at_current_token(
-                    "Enum member expected.",
-                    diagnostic_codes::ENUM_MEMBER_EXPECTED,
-                );
-                self.next_token(); // Skip the comma
-                continue;
-            }
-
-            // Handle @ inside enum body - not a valid enum member start.
-            // Emit TS1132 and break out so the outer statement parser handles the
-            // decorator-like syntax (producing TS1146 + TS1128 matching tsc).
-            if self.is_token(SyntaxKind::AtToken) {
-                self.parse_error_at_current_token(
-                    "Enum member expected.",
-                    diagnostic_codes::ENUM_MEMBER_EXPECTED,
-                );
-                break;
-            }
-
-            // Enum member names can be identifiers, string literals, or computed property names.
-            // Numeric literals are parsed as names for error recovery (TS2452 reported by checker).
-            // Computed property names ([x]) are not valid in enums but we recover gracefully.
-            let name = if self.is_token(SyntaxKind::OpenBracketToken) {
-                // Parse computed property name for recovery. TS1164 is emitted by the
-                // checker (grammar check), not the parser, matching tsc's behavior.
-                // This avoids position-based dedup conflicts with TS1357.
-                self.parse_property_name()
-            } else if self.is_token(SyntaxKind::StringLiteral) {
-                self.parse_string_literal()
-            } else if self.is_token(SyntaxKind::NumericLiteral) {
-                // Parse numeric literal as name for recovery (checker emits TS2452)
-                self.parse_numeric_literal()
-            } else if self.is_token(SyntaxKind::BigIntLiteral) {
-                // Parse bigint literal as name for recovery (checker emits TS2452)
-                self.parse_bigint_literal()
-            } else if self.is_token(SyntaxKind::PrivateIdentifier) {
-                self.parse_error_at_current_token(
-                    "An enum member cannot be named with a private identifier.",
-                    diagnostic_codes::AN_ENUM_MEMBER_CANNOT_BE_NAMED_WITH_A_PRIVATE_IDENTIFIER,
-                );
-                self.parse_private_identifier()
-            } else {
-                self.parse_identifier_name()
-            };
-
-            // Check for unexpected token after enum member name - emit TS1357.
-            // `tsc` still records the malformed member before recovering, so emit
-            // continues to allocate enum values for invalid names such as
-            // `name: 1` and `name;`.
-            if !self.is_token(SyntaxKind::EqualsToken)
-                && !self.is_token(SyntaxKind::CommaToken)
-                && !self.is_token(SyntaxKind::CloseBraceToken)
-                && !self.is_token(SyntaxKind::EndOfFileToken)
-            {
-                self.parse_error_at_current_token(
-                    "An enum member name must be followed by a ',', '=', or '}'.",
-                    diagnostic_codes::AN_ENUM_MEMBER_NAME_MUST_BE_FOLLOWED_BY_A_OR,
-                );
-
-                let member_end = self.arena.get(name).map_or(start_pos, |node| node.end);
-                let member = self.arena.add_enum_member(
-                    syntax_kind_ext::ENUM_MEMBER,
-                    start_pos,
-                    member_end,
-                    EnumMemberData {
-                        name,
-                        initializer: NodeIndex::NONE,
-                    },
-                );
-                members.push(member);
-
-                // Recover by moving past one offending token unless that token
-                // can itself start the next enum member. This keeps namelike
-                // recovery tokens (`any`, `"hello"`, `1`) available to the next
-                // iteration, matching `tsc`'s invalid-member AST.
-                let starts_member = self.is_token(SyntaxKind::OpenBracketToken)
-                    || self.is_token(SyntaxKind::StringLiteral)
-                    || self.is_token(SyntaxKind::NumericLiteral)
-                    || self.is_token(SyntaxKind::BigIntLiteral)
-                    || self.is_token(SyntaxKind::PrivateIdentifier)
-                    || self.is_identifier_or_keyword();
-                if !starts_member {
-                    self.next_token();
-                }
-                continue;
-            }
-
-            let initializer = if self.parse_optional(SyntaxKind::EqualsToken) {
-                self.parse_assignment_expression()
-            } else {
-                NodeIndex::NONE
-            };
-
-            let end_pos = self.token_end();
-            let member = self.arena.add_enum_member(
-                syntax_kind_ext::ENUM_MEMBER,
-                start_pos,
-                end_pos,
-                EnumMemberData { name, initializer },
-            );
-            members.push(member);
-
-            // Parse comma or recover with missing comma
-            if !self.parse_optional(SyntaxKind::CommaToken) {
-                // Recovery: If the next token looks like the start of a valid enum member,
-                // emit TS1357 and continue parsing instead of breaking.
-                // tsc uses TS1357 (enum-specific) rather than generic TS1005 here.
-                if self.is_token(SyntaxKind::Identifier)
-                    || self.is_token(SyntaxKind::StringLiteral)
-                    || self.is_token(SyntaxKind::PrivateIdentifier)
-                    || self.is_token(SyntaxKind::OpenBracketToken)
-                {
-                    self.parse_error_at_current_token(
-                        "An enum member name must be followed by a ',', '=', or '}'.",
-                        diagnostic_codes::AN_ENUM_MEMBER_NAME_MUST_BE_FOLLOWED_BY_A_OR,
-                    );
-                    // Continue to next iteration to parse the next member
-                    continue;
-                }
-                break;
-            }
-        }
-
-        Self::make_node_list(members)
     }
 
     // =========================================================================
